Change the Collabnet Authentication from Soap to REST
[artf422917]

Removed the soap jar that is not used anymore.
Removed the soap related classes.
