+++ README.md
@@ -2,7 +2,7 @@ Older versions of this plugin may not be safe to use. Please review the followin
 
 -   [Plugin globally disables SSL/TLS certification validation in Jenkins](https://jenkins.io/security/advisory/2018-06-25/#SECURITY-941)
 
-The Digital.ai Plugin is an integration with [Digital.ai TeamForge](https://digital.ai/products/teamforge/ "Digital.ai TeamForge"). Don't
+The Digital.ai Plugin is an integration with [Digital.ai TeamForge](https://digital.ai/products/teamforge/ "Digital.ai TeamForge").
 
 Provided features include:
