--- lib/com/collabnet/ce/soap50/webservices/sf_soap50_sdk/1.1/sf_soap50_sdk-1.1.jar.sha1
@@ -1,1 +1,0 @@
-2c491a189974163344ab1607464307982a0d7046
