--- lib/com/collabnet/ce/soap50/webservices/sf_soap50_sdk/1.1/sf_soap50_sdk-1.1.pom
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><project>
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>com.collabnet.ce.soap50.webservices</groupId>
-  <artifactId>sf_soap50_sdk</artifactId>
-  <version>1.1</version>
-</project>
