--- lib/com/collabnet/ce/soap50/webservices/sf_soap50_sdk/1.1/sf_soap50_sdk-1.1.pom.sha1
@@ -1,1 +1,0 @@
-e51d14636171772edc6e2f6bf07a1a960ae89275
