--- lib/com/collabnet/ce/soap50/webservices/sf_soap50_sdk/1.2/sf_soap50_sdk-1.2.jar.md5
@@ -1,1 +1,0 @@
-64c87a0e45fa89eeeda9915ba9c442a9
