--- lib/com/collabnet/ce/soap50/webservices/sf_soap50_sdk/1.2/sf_soap50_sdk-1.2.jar.sha1
@@ -1,1 +1,0 @@
-541191f8fdc495e93b10a423b8c218bdaa941460
