--- lib/com/collabnet/ce/soap50/webservices/sf_soap50_sdk/1.2/sf_soap50_sdk-1.2.pom.md5
@@ -1,1 +1,0 @@
-a56641eced6c2a33c6bad67a58b3ba5c
