--- lib/com/collabnet/ce/soap50/webservices/sf_soap50_sdk/1.2/sf_soap50_sdk-1.2.pom.sha1
@@ -1,1 +1,0 @@
-83f955326e2aadc5fe9a8b8febf909b443332672
