--- lib/com/collabnet/ce/soap50/webservices/sf_soap50_sdk/maven-metadata.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<metadata>
-  <groupId>com.collabnet.ce.soap50.webservices</groupId>
-  <artifactId>sf_soap50_sdk</artifactId>
-  <version>1.1</version>
-  <versioning>
-    <versions>
-      <version>1.1</version>
-      <version>1.2</version>
-    </versions>
-    <lastUpdated>20091207233858</lastUpdated>
-  </versioning>
-</metadata>
