--- lib/com/collabnet/ce/soap50/webservices/sf_soap50_sdk/maven-metadata.xml.md5
@@ -1,1 +1,0 @@
-7d9ebda36bff8b219b4e7205127d6bb1
