--- lib/com/collabnet/ce/soap50/webservices/sf_soap50_sdk/maven-metadata.xml.sha1
@@ -1,1 +1,0 @@
-9de81f0ac9ae83988e48fa38b06d2e6219fa97c1
