--- lib/com/collabnet/ce/soap60/webservices/sf_soap60_sdk/1.1/sf_soap60_sdk-1.1.jar.md5
@@ -1,1 +1,0 @@
-6fdacc131ba0dcd6d06d476cf284eb64
