--- lib/com/collabnet/ce/soap60/webservices/sf_soap60_sdk/1.1/sf_soap60_sdk-1.1.jar.sha1
@@ -1,1 +1,0 @@
-48127607fbcd4d9e31c5a5569fd2c82e49980c8e
