--- lib/com/collabnet/ce/soap60/webservices/sf_soap60_sdk/1.1/sf_soap60_sdk-1.1.pom
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>com.collabnet.ce.soap60.webservices</groupId>
-  <artifactId>sf_soap60_sdk</artifactId>
-  <version>1.1</version>
-</project>
