--- lib/com/collabnet/ce/soap60/webservices/sf_soap60_sdk/1.1/sf_soap60_sdk-1.1.pom.md5
@@ -1,1 +1,0 @@
-380d337f0c2ddb8d2b2b9169f56f0d33
