--- lib/com/collabnet/ce/soap60/webservices/sf_soap60_sdk/1.1/sf_soap60_sdk-1.1.pom.sha1
@@ -1,1 +1,0 @@
-d6910a7f8a8f688606e0b3cfac9a32560c68f996
