--- lib/com/collabnet/ce/soap60/webservices/sf_soap60_sdk/maven-metadata.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<metadata>
-  <groupId>com.collabnet.ce.soap60.webservices</groupId>
-  <artifactId>sf_soap60_sdk</artifactId>
-  <versioning>
-    <release>1.1</release>
-    <versions>
-      <version>1.1</version>
-    </versions>
-    <lastUpdated>20160505114104</lastUpdated>
-  </versioning>
-</metadata>
