--- lib/com/collabnet/ce/soap60/webservices/sf_soap60_sdk/maven-metadata.xml.md5
@@ -1,1 +1,0 @@
-9dbee46ba7c0115ba3b3be892a0becb3
