--- lib/com/collabnet/ce/soap60/webservices/sf_soap60_sdk/maven-metadata.xml.sha1
@@ -1,1 +1,0 @@
-e3fc7b76a50e1ab4b36b5aef166deb26699fca94
