--- lib/com/collabnet/contrib/httpclient-contrib-ssl/1.0/httpclient-contrib-ssl-1.0.jar.sha1
@@ -1,1 +1,0 @@
-f4d52b3a65c2b3bafea2ecbf5165f1809111f358
