--- lib/com/collabnet/contrib/httpclient-contrib-ssl/1.0/httpclient-contrib-ssl-1.0.pom
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><project>
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>com.collabnet.contrib</groupId>
-  <artifactId>httpclient-contrib-ssl</artifactId>
-  <version>1.0</version>
-</project>
