--- lib/com/collabnet/contrib/httpclient-contrib-ssl/1.0/httpclient-contrib-ssl-1.0.pom.sha1
@@ -1,1 +1,0 @@
-f70d05d8e1c2ada85cd2965c488594538bf6282f
