--- lib/com/collabnet/contrib/httpclient-contrib-ssl/maven-metadata.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><metadata>
-  <groupId>com.collabnet.contrib</groupId>
-  <artifactId>httpclient-contrib-ssl</artifactId>
-  <version>1.0</version>
-  <versioning>
-    <versions>
-      <version>1.0</version>
-    </versions>
-    <lastUpdated>20090213232616</lastUpdated>
-  </versioning>
-</metadata>
