--- lib/com/collabnet/contrib/httpclient-contrib-ssl/maven-metadata.xml.md5
@@ -1,1 +1,0 @@
-77be573289f6334260d0f0b80bc2e5b9
