--- lib/com/collabnet/contrib/httpclient-contrib-ssl/maven-metadata.xml.sha1
@@ -1,1 +1,0 @@
-0831a771d776d3f1c141941ee4238a4b751db48b
