--- lib/com/collabnet/cubit/api/cubit-rest-connector/1.0/cubit-rest-connector-1.0.jar.md5
@@ -1,1 +1,0 @@
-7924296025484ac799cb2fad4fcda0d7
