--- lib/com/collabnet/cubit/api/cubit-rest-connector/1.0/cubit-rest-connector-1.0.jar.sha1
@@ -1,1 +1,0 @@
-adbc8966cbd7b1d13bcd4a589dee903837ed9ed3
