--- lib/com/collabnet/cubit/api/cubit-rest-connector/1.0/cubit-rest-connector-1.0.pom
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><project>
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>com.collabnet.cubit.api</groupId>
-  <artifactId>cubit-rest-connector</artifactId>
-  <version>1.0</version>
-</project>
