--- lib/com/collabnet/cubit/api/cubit-rest-connector/1.0/cubit-rest-connector-1.0.pom.md5
@@ -1,1 +1,0 @@
-193d6d3a36837a1af8d296c341ba41c0
