--- lib/com/collabnet/cubit/api/cubit-rest-connector/1.0/cubit-rest-connector-1.0.pom.sha1
@@ -1,1 +1,0 @@
-bd5981c1df1984b575084a218b440f88c1727ebe
