--- lib/com/collabnet/cubit/api/cubit-rest-connector/maven-metadata.xml.md5
@@ -1,1 +1,0 @@
-a9b14ba5c772c48e342dd025818e61f2
