--- lib/com/collabnet/cubit/api/cubit-rest-connector/maven-metadata.xml.sha1
@@ -1,1 +1,0 @@
-f86055c0e6b2b2d5b3dcd5071f9f51ce0a638758
